moved the turn method insid the main menu , the current player helper inside the logic
