+++ lib/game.rb
@@ -1,53 +1,66 @@
 require_relative './helper.rb'
 require_relative './menu.rb'
 require_relative './logic.rb'
-require_relative './player.rb'
 
-class Game 
-    def winner(board)
-        if win(board).class == Array
-          win_player = win(board)
-          win_token = win_player[0]
-          board[win_token]
-        else
-          draw(board)
-        end
-    end
-  
-  def game_end(board)
-    win(board) || draw(board)
+WIN_POSSIBILITY = [
+  [0, 1, 2],
+  [3, 4, 5],
+  [6, 7, 8],
+  [0, 3, 6],
+  [1, 4, 7],
+  [2, 5, 8],
+  [0, 4, 8],
+  [2, 4, 6]
+].freeze
+
+class Game
+  attr_reader :win_posibily
+
+  def initialize(win_posibily = WIN_POSSIBILITY)
+    @win_posibily = win_posibily
   end
 
-  def board_full(board)
-    element = board.all? do |el|
-      if el == 'X'
-        true
-      elsif el == 'O'
-        true
+  def logic
+    @logic = Logic.new
+  end
+
+  def win(board)
+    win_posibily.each do |win_pos|
+      win_pos_one = win_pos[0]
+      win_pos_two = win_pos[1]
+      win_pos_three = win_pos[2]
+
+      pos_one = board[win_pos_one]
+      pos_two = board[win_pos_two]
+      pos_three = board[win_pos_three]
+
+      if pos_one == 'X' && pos_two == 'X' && pos_three == 'X' || pos_one == 'O' && pos_two == 'O' && pos_three == 'O'
+        return win_pos
       end
     end
-    element
+    false
+  end
+
+  def winner(board)
+    if win(board).class == Array
+      win_player = win(board)
+      win_token = win_player[0]
+      board[win_token]
+    else
+      draw(board)
+    end
+  end
+
+  def game_end(board)
+    win(board) || logic.draw(board)
   end
 
   def move(board, index, token = 'X')
     board[index] = token
   end
 
-  
-
-
-  
-def available_slots(board)
-    slots = []
-    board.each do |x|
-      slots << x if x.is_a? Integer
-    end
-    if slots.count.even?
-      puts 'Player 2\'s turn'
-    else
-      puts 'Player 1\'s turn'
-    end
-    slots.each { |x| print x.to_s + ' ' }
+  def current_player(board)
+    count_move(board).even? ? 'X' : 'O'
   end
 
   def count_move(board)
@@ -61,16 +74,4 @@ def available_slots(board)
     end
     counter
   end
-  
-  def postion_taken(board, index)
-    if board[index] == '^[1-9]'
-      false
-    elsif board[index] == 'X'
-      true
-    elsif board[index] == 'O'
-      true
-    end
-  end
-  
-  
 end
